@@ -1,24 +1,39 @@
 {
-  "name": "electron-with-create-react-app",
-  "version": "0.1.0",
-  "private": true,
-  "devDependencies": {
-    "electron": "^1.7.9",
-    "foreman": "^2.0.0",
-    "react-scripts": "0.8.5"
-  },
-  "dependencies": {
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1"
-  },
-  "homepage": "./",
-  "main": "src/electron-starter.js",
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "electron": "electron .",
-    "dev": "nf start -p 3000"
-  }
+    "name": "electron-with-create-react-app",
+    "version": "0.1.0",
+    "private": true,
+    "devDependencies": {
+        "electron": "^1.7.9",
+        "foreman": "^2.0.0",
+        "husky": "^3.0.9",
+        "prettier": "1.18.2",
+        "pretty-quick": "^2.0.0",
+        "react-scripts": "0.8.5"
+    },
+    "dependencies": {
+        "@material-ui/core": "^4.5.1",
+        "aes-js": "^3.1.2",
+        "bootstrap": "^4.3.1",
+        "js-sha512": "^0.8.0",
+        "react": "^16.11.0",
+        "react-bootstrap": "^1.0.0-beta.14",
+        "react-dom": "^16.11.0",
+        "react-dropzone": "^10.1.10",
+        "uuid": "^3.3.3"
+    },
+    "homepage": "./",
+    "main": "src/electron-starter.js",
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test --env=jsdom",
+        "eject": "react-scripts eject",
+        "electron": "electron .",
+        "dev": "nf start -p 3000"
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "pretty-quick --staged"
+        }
+    }
 }
